vm-migration: add vCPU throttling (auto-converge) for pre-copy
## About

auto-converge (vCPU throttling) is a crucial technique to migrate
VMs with a high dirty rate (high working set with intense usage).
It is an alternative to postcopy migration, which is not yet
implemented in Cloud Hypervisor.

## Implementation

vCPU throttling was implemented with a different thread and a
manager for that thread. It is possible to abort vCPU throttling
in case one aborts a live-migration, for example.

The rather complex thread state management is covered in unit
tests ensuring liveliness throughout all possible scenarios.

The throttling itself is implemented by utilizing the CpuManager's
pause() and resume() functions.

## Effective Behaviour

- auto-converging starts always after four memory delta transfer
  iterations (not configurable yet)
- every two iterations, it is increased again
- step size is 10%
- maximum is 99%
- time window for throttling is 100ms. So with 99% throttling,
  the vCPU is paused for 99ms and runs for 1ms, and again.

Signed-off-by: Philipp Schuster <philipp.schuster@cyberus-technology.de>
On-behalf-of: SAP philipp.schuster@sap.com